more fixes.

+++ src/fckey.h
@@ -3,8 +3,8 @@ typedef union {
     uint8_t c[16];
 } fckey_t;
 
-extern int FCKEY_BITS;
-#define FCKEY_MAX ((1L << FCKEY_BITS) - 1)
+extern unsigned int FCKEY_BITS;
+#define FCKEY_MAX ((uint64_t) ((((uint64_t)1L) << FCKEY_BITS) - 1L))
 
 #define FCKEY_FMT "%0.*lo%0.*lo%0.*lo"
 #define FCKEY_PRINT_PREFIX(x, n) \
@@ -25,5 +25,3 @@ void fckey_set_max(fckey_t * key);
 int fckey_is_zero(fckey_t * key);
 void fckey_xor(fckey_t * result, fckey_t * key1, fckey_t * key2);
 void fckey_rightshift(fckey_t * key, int offset);
-
-
